src/main/java/brozart/authorization/client/ClientService.java
package brozart.authorization.client;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.oauth2.provider.ClientDetails;
import org.springframework.security.oauth2.provider.ClientDetailsService;
import org.springframework.security.oauth2.provider.ClientRegistrationException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
public class ClientService implements ClientDetailsService {

    private final ClientRepository clientRepository;

    @Autowired
    public ClientService(final ClientRepository clientRepository) {
        this.clientRepository = clientRepository;
    }

    @Override
    @Transactional(readOnly = true)
    public ClientDetails loadClientByClientId(final String clientId) throws ClientRegistrationException {
        final Client clientById = clientRepository.findByClientId(clientId);
        return new CustomClientDetails(clientById);
    }
}
